Updated EAD parser definitions.
Based on the Dutch National Archive production data, a new parser has
been generated. The support for numbered clevels has been disabled for
now. This should be replaced in the future with a generated version that
is easier to maintain.

The generated parser required many changes into how the data in the
parser was accessed. Also many new objects and attributes have been
added.