Refactor search package to use Vector and be phrase and tag aware.
The tokenizer is aware of pseudo-HTML tags and ignores these for
term-counts (termVectors). Each token is counted with the token.Vector.
Only when a token is a term will the token.TermVector be incremented.

search.Matches will only count Vectors which are a 'term'. Search/Index
implementation should be refactors to use these Vectors instead.