Add test protect against mapping changes that have no migrations.
After the first index has been created changes to the mapping properties
must be added to the mapping migrations otherwise this will result in
indexing errors. Because the hashes of the mapping are now part of the
test suite they should fail when the mapping is changed without updates
to the mappingupdate. The hashes must then be manually updated in the
tests.