Initial version of multi-index bulk.Service.
This will replace the hub3 bulkapi functionality. The current version
has support to write to multiple index types. So both legacy v1 and v2
can be written to at the same time. This is important during this
testing phase especially for the backwards compatible v1 API on the v2
index structure.