Feature/sort fields store proto (#63)
FragmentGraph needs the Protobuf fields in order to store messages in Elasticsearch. This was removed by accident because we do not want this field to show up in the API output. 

The resource.NewFields has a fieldMap map that has unsorted sub maps. But we need to preserve the order of the resource entries in FragmentGraph.Fields.