Dao functionality to update update and process dao-links.
This new functionality allows for updating the daolinks independently
from processing the EAD. The EAD resources are now under version
control but the mets/dao resources not. They are indexed directly.

Some of the dataset generate up to 7m scan records for a single dataset.
This was to much for the current version implementation. That scales
well up to about 500 thousand records.

Orphan control and elasticsearch sync must still be implemented.