Initial version of the Time Revision Store service.
This version works well for smaller datasets less than 250k records.
Some of the larger datasets (up to 8m record) are not suitable for this
solution. The next iteration needs to be better integrated into the
Publisher model. This means that for each dataset the publisher can
subscribe for changes.