Feature/merge mono dev (#153)
* hub3: enable droporphans of lowercase 'findingaid' tag

In some cases this caused the findingAid object not to be dropped from
the index.

* hub3: update changelog

* hub3: docker-compose fuseki writes to local volume.

This fixes the problem that fuseki corrupts when reloading
docker-compose.

* hub3: LDF always returns 'text/turtle'

* hub3: move memory storage package to x

* hub3: refactor elasticsearch storage into driver package

The customClient should wrap all elasticsearch specific functionality.
The goal is to reduce elasticsearch packages being directly used in
various packages.

* hub3: initial version of postgresql service.

* hub3: remove gorm package

* hub3: added multi-tenant configuration support to domain.Organization

* hub3: refactor for multitenancy based on OrgID across all packages

This is a major refactor as preparation for domain.OrganizationConfig
based configuration.

hub3/config will be removed in future commits.

* hub3: added sitemap package

This is an initial version.

* hub3: added dedicade esproxy service.

* hub3: update ikuzoctl configuration objects for domain.Service.

This interface makes dependency injection more predictable.

* hub3: added allowed ports filter to imageproxy

Via the toml via you can specify which ports are allowed. This way you
can prevent port scans via the image proxy.

* hub3: changed resource package name to rdf.

This new name fits better with the intention of the package. It also
makes more sense when writing imports, e.g. rdf.NewGraph.

* hub3: added 'render' helper package.

This uniformly handles how content in written to http responses.
Also handling of error responses is streamlined.

* hub3: added basic support for extracting text from alto files.

* hub3: use ikuzo/render package in legacy handlers

This also enables sentry support.

* hub3: fixed order for processing triples in ead processing.

This fix will allow for triples to be added in the CreateTree function
and still be processed in the FragmentGraph.

* hub3: added support for rdf.Graph generation from FragmentResources.

This is part of the migration to rdf.Record.

* hub3: deprecate index.Client in favor of elasticsearch driver Client.

* hub3: update organization_config configuration

This is part of the migration away from config/config.go global config.

* hub3: added support for sentry.io error aggregation.

This can be configured through the hub3.toml configuration.

For now it is a per hub error aggregator

* hub3: added support for resource based indexing to rdf.Graph

This internal indexing can be used by serialization formats and for
queries direcly on the graph.

Also pointers to IRI have been removed. This was a source of confusion
and made the usage more difficult than needed.

* hub3: Initial support for mappingxml  RDF serialization format.

This format takes a group of triples and turns it into mappable XML. The
main purpose is to provide the Sip-Creator with a mappable format for
raw RDF data.

* hub3: bulk sparql updates support filtered by urn:private.

This must later be replaced by rdf.Graph natively.

* hub3: added more default namespaces to namespace package.

Also added additional tests. There seems to be a bug when multiple
prefixes are added from the map.

In the future, must add support for preferred prefixes.

* hub3: migrate es search statistics to own sub-package.

* hub3: initial version of generic lod resolver package.

The lod.Resolver interface can be implemented by various backends to
resolve rdf.Subject uris. This interface must take take of the
multitenancy.

* hub3: changed sitemap package to support stores.

This commit also contains the initial implementation of an elasticsearch
based sitemap store. The file package will also contain a sitemap based
on the RecordIndex which will be the default in the future.

The elasticsearch package will be used for dynamic sitemaps.

* hub3: added initial support for harvest.Syncer interface.

This interface should wrap other harvesters such as oaipmh, sparql, etc.

This harvester has internal test for an SRU interface.

* hub3: changed oai-pmh service to support stores.

The previous oaipmh service has been moved to the harvest sub-package.
This was mainly geared towards supporting syncing EAD.

* hub3: initial support for file package.

This is an disk-based implementation for various hub3 stores such as
sitemap, resolve, oai-pmh, etc.

It will also support direct uploads of delimited nquads from narthex.

* hub3: initial implementation of sparql service

This should replace all the sparql queries and calls in individual
packages.

* hub3: fixed predictable ordering for namespace defaults.

Before the namespace defaults were structured in a map, this caused
unpredicatble behaviour in tests and API.

* hub3: fixes for unit tests and ikuzo server after refactor.

* hub3: update changelog

* Add tls support for LDF hypermedia pagination.